backend/requirements.txt
SQLAlchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn==0.31.0
fastapi==0.115.0
email_validator==2.2.0
pydantic==1.10.13
pydantic_core==2.16.3
google-auth==2.35.0
requests==2.31.0
google-auth-oauthlib==1.2.0
python-jose[cryptography]==3.3.0